# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateMemberRequest": {
        "additionalProperties": false,
        "properties": {
          "display_name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "maxLength": 320,
            "title": "Email",
            "type": "string"
          },
          "password": {
            "description": "Initial password (12 to 128 characters); share it securely.",
            "maxLength": 1024,
            "title": "Password",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role",
            "default": "MEMBER"
          }
        },
        "required": [
          "email",
          "display_name",
          "password"
        ],
        "title": "CreateMemberRequest",
        "type": "object"
      },
...
      "DataResponse_Member_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Member"
          }
        },
        "required": [
          "data"
        ],
        "title": "DataResponse[Member]",
        "type": "object"
      },
...
      "DataResponse_TenantDetail_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TenantDetail"
          }
        },
        "required": [
          "data"
        ],
        "title": "DataResponse[TenantDetail]",
        "type": "object"
      },
...
      "DataResponse_list_Member__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Member"
            },
            "title": "Data",
            "type": "array"
          }
        },
        "required": [
          "data"
        ],
        "title": "DataResponse[list[Member]]",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "description": "Stable machine-readable error code.",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Same value as the X-Request-ID response header.",
            "title": "Request Id"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorDetail": {
        "properties": {
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "Member": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "last_login_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "required": [
          "id",
          "email",
          "display_name",
          "role",
          "is_active",
          "created_at",
          "last_login_at"
        ],
        "title": "Member",
        "type": "object"
      },
...
      "Role": {
        "enum": [
          "OWNER",
          "ADMIN",
          "MEMBER"
        ],
        "title": "Role",
        "type": "string"
      },
...
      "TenantDetail": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "member_count": {
            "title": "Member Count",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "created_at",
          "member_count"
        ],
        "title": "TenantDetail",
        "type": "object"
      },
...
      "UpdateMemberRequest": {
        "additionalProperties": false,
        "properties": {
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "false signs the user out everywhere and blocks sign-in.",
            "title": "Is Active"
          },
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Role"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "UpdateMemberRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nRead-only analytics API over YouTube trending snapshots (PostgreSQL `app` schema) and a frozen ML model.\n\n**Metric semantics**\n* *Latest snapshot* = each video's latest snapshot **within the filtered dataset** (date range,\n  countries, categories), not its globally latest snapshot.\n* **Snapshot-level** metrics count trending appearances (country x video x day), e.g. trending volume.\n* **Video-level** metrics count each video once via its latest filtered snapshot, e.g. unique videos,\n  views, engagement rate.\n* **Country-level** metrics use each video's latest filtered snapshot in that country. Country\n  `views_of_trending_videos` are **global lifetime views** of videos that trended there; they are\n  **not additive** across countries.\n* View counts are lifetime totals: historical snapshots are **never summed** for views.\n* Period change: counts and views in **percent**; engagement rate in **percentage points**.\n\n**Authentication**: sign in via `/api/v1/auth/login` (or register). The session is an HttpOnly\ncookie; every `/api/v1` endpoint except register/login requires it. Analytics endpoints need\n`VIEW_ANALYTICS`, predictions `MAKE_PREDICTION`, member management `MANAGE_USERS`. The YouTube\ndataset is global and read-only; workspaces (tenants) own only their users and memberships.\nHealth probes are public.\n",
    "title": "YouTube Trending Intelligence API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/tenant": {
      "get": {
        "operationId": "get_tenant_api_v1_tenant_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_TenantDetail_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not signed in or session expired."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing permission or role rule violated."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request."
          }
        },
        "summary": "Your workspace",
        "tags": [
          "tenant"
        ]
      }
    },
    "/api/v1/tenant/members": {
      "get": {
        "operationId": "list_members_api_v1_tenant_members_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_list_Member__"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not signed in or session expired."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing permission or role rule violated."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request."
          }
        },
        "summary": "Members of your workspace (MANAGE_USERS)",
        "tags": [
          "tenant"
        ]
      },
      "post": {
        "operationId": "create_member_api_v1_tenant_members_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMemberRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_Member_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not signed in or session expired."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing permission or role rule violated."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Email already registered."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request."
          }
        },
        "summary": "Add a user to your workspace with an initial password (MANAGE_USERS)",
        "tags": [
          "tenant"
        ]
      }
    },
    "/api/v1/tenant/members/{user_id}": {
      "patch": {
        "operationId": "update_member_api_v1_tenant_members__user_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_Member_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not signed in or session expired."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing permission or role rule violated."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such member in your workspace."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Would leave no active owner."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request."
          }
        },
        "summary": "Change a member's role or active status (MANAGE_USERS)",
        "tags": [
          "tenant"
        ]
      }
    },
...
  }
}